add traffic logging / create access list (#13)
* log all HTTP requests
* remove traceback from log for CustomErrors 
* update to have CustomErrors with status_code
* update docker script to auto-update before launch
* fix low security risk on headers
* add ip filtering for bots
* update templates parameters
* add admin page to manage access list

+++ docker_build_run.sh
@@ -6,6 +6,10 @@ function has_docker_compose() {
   fi
 }
 
+function get_running_pid() {
+  PID=$(ps -ef | grep "uvicorn app.main" | grep -v grep | awk '{print $2}')
+}
+
 if [ -z "$(docker images -q cecc-base:latest)" ]; then
   echo ">>> Creating cecc-base image"
   if has_docker_compose; then
@@ -17,6 +21,18 @@ fi
 
 if [ "$1" == "update" ] && [ -n "$(docker images -q cecc:latest)" ]; then
   echo ">>> Removing cecc image"
+
+  get_running_pid
+  if [ -z "$PID" ]; then
+    echo "    - 'uvicorn app.main' not running."
+  else
+    sudo kill -9 $PID
+    echo "    - 'uvicorn app.main' with PID $PID has been killed."
+  fi
+
+  git pull
+
+  docker rm -f cecc
   docker rmi cecc || { echo "Failed to remove app image"; exit 1; }
 fi
 
@@ -29,8 +45,14 @@ if [ -z "$(docker images -q cecc:latest)" ]; then
   fi
 fi
 
-if has_docker_compose; then
-  docker-compose run --rm --service-ports app
-else
-  docker run --rm -p 8443:443 -v ./data:/cecc/data cecc
+get_running_pid
+if [ -z "$PID" ]; then
+  if has_docker_compose; then
+    docker-compose run --rm --service-ports app
+  else
+    nohup docker run --rm -p 8443:443 -v ./data:/cecc/data cecc &
+  fi
 fi
+
+VERSION=$(curl --silent --insecure https://127.0.0.1:8443/health)
+echo $VERSION
